requirements.txt
python-telegram-bot
python-dotenv==0.15.0
requests
pillow